Split components kits provider into separate file
