+++ packages/app/ui/contexts/componentsKits/componentsKits.tsx
@@ -2,39 +2,12 @@ import { Story } from '@tloncorp/api/urbit';
 import {
   CollectionRendererId,
   DraftInputId,
-  JSONValue,
   PostContentRendererId,
 } from '@tloncorp/shared';
 import * as db from '@tloncorp/shared/db';
 import { ReactElement, createContext, useContext } from 'react';
-import { Text } from 'react-native';
 
-import { AudioPost } from '../../components/AudioPost';
-import { ChatMessage } from '../../components/ChatMessage';
-import { ColorPost } from '../../components/ColorPost';
-import { useContactName } from '../../components/ContactNameV2';
-import { EditableNotePostContent } from '../../components/EditableNotePostContent';
-import { GalleryPost } from '../../components/GalleryPost';
-import { NotebookPost } from '../../components/NotebookPost';
-import { YellPost } from '../../components/YellPost';
-import {
-  ChatInput,
-  ColorInput,
-  DraftInputContext,
-  GalleryInput,
-  MicInput,
-  NotebookInput,
-} from '../../components/draftInputs';
-import { ButtonInput } from '../../components/draftInputs/ButtonInput';
-import { BoardroomPostCollectionView } from '../../components/postCollectionViews/BoardroomPostCollectionView';
-import {
-  CardsPostCollection,
-  SingleCardPostCollection,
-} from '../../components/postCollectionViews/CardsPostCollectionView';
-import { CarouselPostCollection } from '../../components/postCollectionViews/CarouselPostCollectionView';
-import { ListPostCollection } from '../../components/postCollectionViews/ListPostCollectionView';
-import { StrobePostCollectionView } from '../../components/postCollectionViews/StrobePostCollectionView';
-import { PostSummaryCollectionView } from '../../components/postCollectionViews/SummaryCollectionView';
+import { DraftInputContext } from '../../components/draftInputs';
 import { IPostCollectionView } from '../../components/postCollectionViews/shared';
 
 type RenderItemProps = {
@@ -85,11 +58,11 @@ export type MinimalRenderItemProps = {
 };
 export type MinimalRenderItemType = React.ComponentType<MinimalRenderItemProps>;
 
-type DraftInputRendererComponent = React.ComponentType<{
+export type DraftInputRendererComponent = React.ComponentType<{
   draftInputContext: DraftInputContext;
 }>;
 
-interface ComponentsKitContextValue {
+export interface ComponentsKitContextValue {
   collectionRenderers: Readonly<
     Partial<{ [Id in CollectionRendererId]: IPostCollectionView }>
   >;
@@ -101,72 +74,11 @@ interface ComponentsKitContextValue {
   >;
 }
 
-const BUILTIN_CONTENT_RENDERERS: { [id: string]: RenderItemType } = {
-  [PostContentRendererId.chat]: ChatMessage,
-  [PostContentRendererId.gallery]: GalleryPost,
-  [PostContentRendererId.notebook]: NotebookPost,
-  [PostContentRendererId.audio]: AudioPost,
-  [PostContentRendererId.color]: ColorPost,
-  [PostContentRendererId.raw]: ({ post, contentRendererConfiguration }) => {
-    const contactName = useContactName(post.author!.id);
-    return (
-      <Text
-        style={{
-          fontFamily: JSONValue.asString(
-            contentRendererConfiguration?.fontFamily,
-            undefined
-          ),
-        }}
-      >
-        {contactName}: {JSON.stringify(post.content)}
-      </Text>
-    );
-  },
-  [PostContentRendererId.yell]: YellPost,
-  [PostContentRendererId.scratchpad]: EditableNotePostContent,
-};
-const BUILTIN_DRAFT_INPUTS: { [id: string]: DraftInputRendererComponent } = {
-  [DraftInputId.chat]: ChatInput,
-  [DraftInputId.gallery]: GalleryInput,
-  [DraftInputId.notebook]: NotebookInput,
-  [DraftInputId.yo]: ({ draftInputContext }) => (
-    <ButtonInput
-      draftInputContext={draftInputContext}
-      messageText={JSONValue.asString(
-        draftInputContext.configuration?.text,
-        'Yo'
-      )}
-      labelText={JSONValue.asString(
-        draftInputContext.configuration?.text,
-        'Yo'
-      )}
-    />
-  ),
-  [DraftInputId.mic]: MicInput,
-  [DraftInputId.color]: ColorInput,
-};
-const BUILTIN_COLLECTION_RENDERERS: {
-  [id in CollectionRendererId]: IPostCollectionView;
-} = {
-  [CollectionRendererId.chat]: ListPostCollection,
-  [CollectionRendererId.gallery]: ListPostCollection,
-  [CollectionRendererId.notebook]: ListPostCollection,
-  [CollectionRendererId.carousel]: CarouselPostCollection,
-  [CollectionRendererId.cards]: CardsPostCollection,
-  [CollectionRendererId.sign]: SingleCardPostCollection,
-  [CollectionRendererId.boardroom]: BoardroomPostCollectionView,
-  [CollectionRendererId.strobe]: StrobePostCollectionView,
-  [CollectionRendererId.summaries]: PostSummaryCollectionView,
-};
-
-const _globalContextValue: ComponentsKitContextValue = {
-  collectionRenderers: BUILTIN_COLLECTION_RENDERERS,
-  inputs: BUILTIN_DRAFT_INPUTS,
-  renderers: BUILTIN_CONTENT_RENDERERS,
-};
-
-const ComponentsKitContext =
-  createContext<ComponentsKitContextValue>(_globalContextValue);
+export const ComponentsKitContext = createContext<ComponentsKitContextValue>({
+  collectionRenderers: {},
+  inputs: {},
+  renderers: {},
+});
 
 export function useComponentsKitContext() {
   return useContext(ComponentsKitContext);
